no-issue: Move 'runCommand' utils into separate file
